requirements.txt
streamlit>=1.38.0
reportlab>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0
dataclasses;python_version<"3.7"
PyYAML>=6.0.0
pdfplumber
